backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
google-generativeai
datasets
sqlalchemy
python-dotenv
pypdf
langchain-text-splitters
